Make WALLS_RADIUS a u8 since it'll mostly be used in relation to room x/y which we generally pass around as u8

+++ src/constants/seasonal.rs
@@ -16,7 +16,7 @@ pub mod season_1 {
     ///
     /// [`ScoreCollector`]: crate::objects::ScoreCollector
     /// [`StructureWall`]: crate::objects::StructureWall
-    pub const WALLS_RADIUS: u32 = 5;
+    pub const WALLS_RADIUS: u8 = 5;
     /// The percentage chance that a given room will have a [`ScoreContainer`]
     /// spawned in it every [`SCORE_CONTAINER_SPAWN_INTERVAL_TICKS`] ticks.
     ///
